Add descriptions to all classes and consolidate
