+++ src/interactive-window/commands/commandRegistry.ts
@@ -4,14 +4,31 @@
 'use strict';
 
 import { inject, injectable, optional } from 'inversify';
-import { CodeLens, ConfigurationTarget, env, Range, Uri, commands } from 'vscode';
-import { IKernelProvider } from '../../kernels/types';
+import {
+    CodeLens,
+    ConfigurationTarget,
+    env,
+    Range,
+    Uri,
+    commands,
+    NotebookCell,
+    NotebookEdit,
+    NotebookRange,
+    Selection,
+    Position,
+    ViewColumn,
+    workspace,
+    WorkspaceEdit
+} from 'vscode';
+import { IKernelProvider, KernelConnectionMetadata } from '../../kernels/types';
 import { ICommandNameArgumentTypeMapping } from '../../commands';
 import {
     IApplicationShell,
+    IClipboard,
     ICommandManager,
     IDebugService,
     IDocumentManager,
+    IVSCodeNotebook,
     IWorkspaceService
 } from '../../platform/common/application/types';
 
@@ -19,18 +36,39 @@ import { IConfigurationService, IDisposable, IDisposableRegistry } from '../../p
 import { DataScience } from '../../platform/common/utils/localize';
 import { isUri, noop } from '../../platform/common/utils/misc';
 import { captureTelemetry } from '../../telemetry';
-import { Commands, Telemetry } from '../../platform/common/constants';
+import { Commands, JVSC_EXTENSION_ID, PYTHON_LANGUAGE, Telemetry } from '../../platform/common/constants';
 import { IDataScienceCodeLensProvider, ICodeWatcher } from '../editor-integration/types';
 import { IInteractiveWindowProvider } from '../types';
 import * as urlPath from '../../platform/vscode-path/resources';
-import { getFilePath } from '../../platform/common/platform/fs-paths';
+import { getDisplayPath, getFilePath } from '../../platform/common/platform/fs-paths';
 import { IExtensionSingleActivationService } from '../../platform/activation/types';
+import { chainWithPendingUpdates } from '../../kernels/execution/notebookUpdater';
+import { ExportFormat, IExportDialog, IFileConverter } from '../../notebooks/export/types';
+import { openAndShowNotebook } from '../../platform/common/utils/notebooks';
+import { JupyterInstallError } from '../../platform/errors/jupyterInstallError';
+import { traceError, traceInfo } from '../../platform/logging';
+import { CommandSource } from '../../platform/testing/common/constants';
+import { generateCellsFromDocument } from '../editor-integration/cellFactory';
+import { IDataScienceErrorHandler } from '../../kernels/errors/types';
+import { INotebookEditorProvider } from '../../notebooks/types';
+import { INotebookExporter, IJupyterExecution } from '../../kernels/jupyter/types';
+import { IFileSystem } from '../../platform/common/platform/types';
+import { IControllerPreferredService } from '../../notebooks/controllers/types';
+import { IStatusProvider } from '../../platform/progress/types';
 
+/**
+ * Class that registers command handlers for interactive window commands.
+ */
 @injectable()
 export class CommandRegistry implements IDisposable, IExtensionSingleActivationService {
     constructor(
         @inject(IDisposableRegistry) private readonly disposables: IDisposableRegistry,
+        @inject(INotebookExporter) @optional() private jupyterExporter: INotebookExporter | undefined,
+        @inject(IJupyterExecution) private jupyterExecution: IJupyterExecution,
         @inject(IDocumentManager) private documentManager: IDocumentManager,
+        @inject(IApplicationShell) private applicationShell: IApplicationShell,
+        @inject(IFileSystem) private fileSystem: IFileSystem,
+        @inject(IConfigurationService) private configuration: IConfigurationService,
         @inject(IDataScienceCodeLensProvider)
         @optional()
         private dataScienceCodeLensProvider: IDataScienceCodeLensProvider | undefined,
@@ -40,9 +78,16 @@ export class CommandRegistry implements IDisposable, IExtensionSingleActivationS
         @inject(IApplicationShell) private appShell: IApplicationShell,
         @inject(IWorkspaceService) private readonly workspace: IWorkspaceService,
         @inject(IInteractiveWindowProvider)
-        @optional()
-        private readonly interactiveWindowProvider: IInteractiveWindowProvider | undefined,
-        @inject(IKernelProvider) private readonly kernelProvider: IKernelProvider
+        private readonly interactiveWindowProvider: IInteractiveWindowProvider,
+        @inject(IKernelProvider) private readonly kernelProvider: IKernelProvider,
+        @inject(IDataScienceErrorHandler) private dataScienceErrorHandler: IDataScienceErrorHandler,
+        @inject(INotebookEditorProvider) protected ipynbProvider: INotebookEditorProvider,
+        @inject(IFileConverter) private fileConverter: IFileConverter,
+        @inject(IExportDialog) private exportDialog: IExportDialog,
+        @inject(IClipboard) private clipboard: IClipboard,
+        @inject(IVSCodeNotebook) private notebook: IVSCodeNotebook,
+        @inject(IControllerPreferredService) private controllerPreferredService: IControllerPreferredService,
+        @inject(IStatusProvider) private statusProvider: IStatusProvider
     ) {
         if (!this.workspace.isTrusted) {
             this.workspace.onDidGrantWorkspaceTrust(this.registerCommandsIfTrusted, this, this.disposables);
@@ -73,6 +118,85 @@ export class CommandRegistry implements IDisposable, IExtensionSingleActivationS
             Commands.EnableLoadingWidgetsFrom3rdPartySource,
             this.enableLoadingWidgetScriptsFromThirdParty
         );
+        this.registerCommand(Commands.CreateNewInteractive, (connection?: KernelConnectionMetadata) =>
+            this.createNewInteractiveWindow(connection)
+        );
+        this.registerCommand(
+            Commands.ImportNotebook,
+            (file?: Uri, _cmdSource: CommandSource = CommandSource.commandPalette) => {
+                return this.listenForErrors(() => {
+                    if (file) {
+                        return this.importNotebookOnFile(file);
+                    } else {
+                        return this.importNotebook();
+                    }
+                });
+            }
+        );
+        this.registerCommand(
+            Commands.ImportNotebookFile,
+            (file?: Uri, _cmdSource: CommandSource = CommandSource.commandPalette) => {
+                return this.listenForErrors(() => {
+                    if (file) {
+                        return this.importNotebookOnFile(file);
+                    } else {
+                        return this.importNotebook();
+                    }
+                });
+            }
+        );
+        this.commandManager.registerCommand(
+            Commands.ExportFileAsNotebook,
+            (file?: Uri, _cmdSource: CommandSource = CommandSource.commandPalette) => {
+                return this.listenForErrors(() => {
+                    if (file) {
+                        return this.exportFile(file);
+                    } else {
+                        const activeEditor = this.documentManager.activeTextEditor;
+                        if (activeEditor && activeEditor.document.languageId === PYTHON_LANGUAGE) {
+                            return this.exportFile(activeEditor.document.uri);
+                        }
+                    }
+
+                    return Promise.resolve();
+                });
+            }
+        );
+        this.registerCommand(
+            Commands.ExportFileAndOutputAsNotebook,
+            (file: Uri, _cmdSource: CommandSource = CommandSource.commandPalette) => {
+                return this.listenForErrors(() => {
+                    if (file) {
+                        return this.exportFileAndOutput(file);
+                    } else {
+                        const activeEditor = this.documentManager.activeTextEditor;
+                        if (activeEditor && activeEditor.document.languageId === PYTHON_LANGUAGE) {
+                            return this.exportFileAndOutput(activeEditor.document.uri);
+                        }
+                    }
+                    return Promise.resolve();
+                });
+            }
+        );
+        this.registerCommand(Commands.ExpandAllCells, async (context?: { notebookEditor: { notebookUri: Uri } }) =>
+            this.expandAllCells(context?.notebookEditor?.notebookUri)
+        );
+        this.registerCommand(Commands.CollapseAllCells, async (context?: { notebookEditor: { notebookUri: Uri } }) =>
+            this.collapseAllCells(context?.notebookEditor?.notebookUri)
+        );
+        this.registerCommand(Commands.ExportOutputAsNotebook, () => this.exportCells());
+        this.registerCommand(
+            Commands.InteractiveExportAsNotebook,
+            (context?: { notebookEditor: { notebookUri: Uri } }) => this.export(context?.notebookEditor?.notebookUri)
+        );
+        this.registerCommand(Commands.InteractiveExportAs, (context?: { notebookEditor: { notebookUri: Uri } }) =>
+            this.exportAs(context?.notebookEditor?.notebookUri)
+        );
+        this.registerCommand(Commands.ScrollToCell, (file: Uri, id: string) => this.scrollToCell(file, id));
+        this.registerCommand(Commands.InteractiveClearAll, this.clearAllCellsInInteractiveWindow);
+        this.registerCommand(Commands.InteractiveRemoveCell, this.removeCellInInteractiveWindow);
+        this.registerCommand(Commands.InteractiveGoToCode, this.goToCodeInInteractiveWindow);
+        this.commandManager.registerCommand(Commands.InteractiveCopyCell, this.copyCellInInteractiveWindow);
     }
     public dispose() {
         this.disposables.forEach((d) => d.dispose());
@@ -483,5 +607,302 @@ export class CommandRegistry implements IDisposable, IExtensionSingleActivationS
     // Open the VS Code outline view
     private async openOutlineView(): Promise<void> {
         return this.commandManager.executeCommand('outline.focus');
+    }
+
+    /* eslint-disable @typescript-eslint/no-explicit-any */
+    private async listenForErrors(promise: () => Promise<any>): Promise<any> {
+        let result: any;
+        try {
+            result = await promise();
+            return result;
+        } catch (err) {
+            traceError('listenForErrors', err as any);
+            this.dataScienceErrorHandler.handleError(err).then(noop, noop);
+        }
+        return result;
+    }
+
+    @captureTelemetry(Telemetry.ExportPythonFileInteractive, undefined, false)
+    private async exportFile(file: Uri): Promise<void> {
+        const filePath = getFilePath(file);
+        if (filePath && filePath.length > 0 && this.jupyterExporter) {
+            // If the current file is the active editor, then generate cells from the document.
+            const activeEditor = this.documentManager.activeTextEditor;
+            if (activeEditor && this.fileSystem.arePathsSame(activeEditor.document.uri, file)) {
+                const cells = generateCellsFromDocument(
+                    activeEditor.document,
+                    this.configuration.getSettings(activeEditor.document.uri)
+                );
+                if (cells) {
+                    // Bring up the export dialog box
+                    const uri = await this.exportDialog.showDialog(ExportFormat.ipynb, file);
+                    await this.waitForStatus(
+                        async () => {
+                            if (uri) {
+                                const notebook = await this.jupyterExporter?.translateToNotebook(cells);
+                                await this.fileSystem.writeFile(uri, JSON.stringify(notebook));
+                            }
+                        },
+                        DataScience.exportingFormat(),
+                        getDisplayPath(file)
+                    );
+                    // When all done, show a notice that it completed.
+                    if (uri && filePath) {
+                        const openQuestion1 = DataScience.exportOpenQuestion1();
+                        const selection = await this.applicationShell.showInformationMessage(
+                            DataScience.exportDialogComplete().format(getDisplayPath(file)),
+                            openQuestion1
+                        );
+                        if (selection === openQuestion1) {
+                            await openAndShowNotebook(uri);
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    @captureTelemetry(Telemetry.ExportPythonFileAndOutputInteractive, undefined, false)
+    private async exportFileAndOutput(file: Uri): Promise<Uri | undefined> {
+        const filePath = getFilePath(file);
+        if (
+            filePath &&
+            filePath.length > 0 &&
+            this.jupyterExporter &&
+            (await this.jupyterExecution.isNotebookSupported())
+        ) {
+            // If the current file is the active editor, then generate cells from the document.
+            const activeEditor = this.documentManager.activeTextEditor;
+            if (
+                activeEditor &&
+                activeEditor.document &&
+                this.fileSystem.arePathsSame(activeEditor.document.uri, file)
+            ) {
+                const cells = generateCellsFromDocument(
+                    activeEditor.document,
+                    this.configuration.getSettings(activeEditor.document.uri)
+                );
+                if (cells) {
+                    // Bring up the export dialog box
+                    const uri = await this.exportDialog.showDialog(ExportFormat.ipynb, file);
+                    if (!uri) {
+                        return;
+                    }
+                    await this.waitForStatus(
+                        async () => {
+                            if (uri) {
+                                const notebook = await this.jupyterExporter?.translateToNotebook(cells);
+                                await this.fileSystem.writeFile(uri, JSON.stringify(notebook));
+                            }
+                        },
+                        DataScience.exportingFormat(),
+                        getDisplayPath(file)
+                    );
+                    // Next open this notebook & execute it.
+                    const editor = await this.notebook
+                        .openNotebookDocument(uri)
+                        .then((document) => this.notebook.showNotebookDocument(document));
+                    const { controller } = await this.controllerPreferredService.computePreferred(editor.notebook);
+                    if (controller) {
+                        await this.commandManager.executeCommand('notebook.selectKernel', {
+                            id: controller.id,
+                            extension: JVSC_EXTENSION_ID
+                        });
+                    }
+                    await this.commandManager.executeCommand('notebook.execute');
+                    return uri;
+                }
+            }
+        } else {
+            await this.dataScienceErrorHandler.handleError(
+                new JupyterInstallError(
+                    DataScience.jupyterNotSupported().format(await this.jupyterExecution.getNotebookError())
+                )
+            );
+        }
+    }
+
+    private async expandAllCells(uri?: Uri) {
+        const interactiveWindow = this.getTargetInteractiveWindow(uri);
+        traceInfo(`Expanding all cells in interactive window with uri ${interactiveWindow?.notebookUri}`);
+        if (interactiveWindow) {
+            await interactiveWindow.expandAllCells();
+        }
+    }
+
+    private async collapseAllCells(uri?: Uri) {
+        const interactiveWindow = this.getTargetInteractiveWindow(uri);
+        traceInfo(`Collapsing all cells in interactive window with uri ${interactiveWindow?.notebookUri}`);
+        if (interactiveWindow) {
+            await interactiveWindow.collapseAllCells();
+        }
+    }
+
+    private exportCells() {
+        const interactiveWindow = this.interactiveWindowProvider?.activeWindow;
+        if (interactiveWindow) {
+            interactiveWindow.export();
+        }
+    }
+
+    private exportAs(uri?: Uri) {
+        const interactiveWindow = this.getTargetInteractiveWindow(uri);
+        if (interactiveWindow) {
+            interactiveWindow.exportAs();
+        }
+    }
+
+    private export(uri?: Uri) {
+        const interactiveWindow = this.getTargetInteractiveWindow(uri);
+        if (interactiveWindow) {
+            interactiveWindow.export();
+        }
+    }
+
+    @captureTelemetry(Telemetry.CreateNewInteractive, undefined, false)
+    private async createNewInteractiveWindow(connection?: KernelConnectionMetadata): Promise<void> {
+        await this.interactiveWindowProvider?.getOrCreate(undefined, connection);
+    }
+
+    private waitForStatus<T>(
+        promise: () => Promise<T>,
+        format: string,
+        file?: string,
+        canceled?: () => void
+    ): Promise<T> {
+        const message = file ? format.format(file) : format;
+        return this.statusProvider.waitWithStatus(promise, message, undefined, canceled);
+    }
+
+    @captureTelemetry(Telemetry.ImportNotebook, { scope: 'command' }, false)
+    private async importNotebook(): Promise<void> {
+        const filtersKey = DataScience.importDialogFilter();
+        const filtersObject: { [name: string]: string[] } = {};
+        filtersObject[filtersKey] = ['ipynb'];
+
+        const uris = await this.applicationShell.showOpenDialog({
+            openLabel: DataScience.importDialogTitle(),
+            filters: filtersObject
+        });
+
+        if (uris && uris.length > 0) {
+            // Don't call the other overload as we'll end up with double telemetry.
+            await this.waitForStatus(
+                async () => {
+                    await this.fileConverter.importIpynb(uris[0]);
+                },
+                DataScience.importingFormat(),
+                getDisplayPath(uris[0])
+            );
+        }
+    }
+
+    @captureTelemetry(Telemetry.ImportNotebook, { scope: 'file' }, false)
+    private async importNotebookOnFile(file: Uri): Promise<void> {
+        const filepath = getFilePath(file);
+        if (filepath && filepath.length > 0) {
+            await this.waitForStatus(
+                async () => {
+                    await this.fileConverter.importIpynb(file);
+                },
+                DataScience.importingFormat(),
+                getDisplayPath(file)
+            );
+        }
+    }
+
+    private async scrollToCell(file: Uri, id: string): Promise<void> {
+        if (id && file) {
+            // Find the interactive windows that have this file as a submitter
+            const possibles = this.interactiveWindowProvider.windows.filter(
+                (w) => w.submitters.findIndex((s) => this.fileSystem.arePathsSame(s, file)) >= 0
+            );
+
+            // Scroll to cell in the one that has the cell. We need this so
+            // we don't activate all of them.
+            // eslint-disable-next-line @typescript-eslint/prefer-for-of
+            for (let i = 0; i < possibles.length; i += 1) {
+                if (await possibles[i].hasCell(id)) {
+                    possibles[i].scrollToCell(id);
+                    break;
+                }
+            }
+        }
+    }
+
+    private async clearAllCellsInInteractiveWindow(context?: { notebookEditor: { notebookUri: Uri } }): Promise<void> {
+        const uri = this.getTargetInteractiveWindow(context?.notebookEditor?.notebookUri)?.notebookUri;
+        if (!uri) {
+            return;
+        }
+
+        // Look for the matching notebook document to add cells to
+        const document = workspace.notebookDocuments.find((document) => document.uri.toString() === uri.toString());
+        if (!document) {
+            return;
+        }
+
+        // Remove the cells from the matching notebook document
+        const edit = new WorkspaceEdit();
+        const nbEdit = NotebookEdit.deleteCells(new NotebookRange(0, document.cellCount));
+        edit.set(document.uri, [nbEdit]);
+        await workspace.applyEdit(edit);
+    }
+
+    private async removeCellInInteractiveWindow(context?: NotebookCell) {
+        const interactiveWindow = this.interactiveWindowProvider.getActiveOrAssociatedInteractiveWindow();
+        const ranges =
+            context === undefined
+                ? interactiveWindow?.notebookEditor?.selections
+                : [new NotebookRange(context.index, context.index + 1)];
+        const document = context === undefined ? interactiveWindow?.notebookEditor?.notebook : context.notebook;
+
+        if (ranges !== undefined && document !== undefined) {
+            await chainWithPendingUpdates(document, (edit) => {
+                ranges.forEach((range) => {
+                    const nbEdit = NotebookEdit.deleteCells(range);
+                    edit.set(document.uri, [nbEdit]);
+                });
+            });
+        }
+    }
+
+    private async goToCodeInInteractiveWindow(context?: NotebookCell) {
+        if (context && context.metadata?.interactive) {
+            const uri = Uri.parse(context.metadata.interactive.uristring);
+            const line = context.metadata.interactive.lineIndex;
+
+            const editor = await this.documentManager.showTextDocument(uri, { viewColumn: ViewColumn.One });
+
+            // If we found the editor change its selection
+            if (editor) {
+                editor.revealRange(new Range(line, 0, line, 0));
+                editor.selection = new Selection(new Position(line, 0), new Position(line, 0));
+            }
+        }
+    }
+
+    private async copyCellInInteractiveWindow(context?: NotebookCell) {
+        if (context) {
+            const settings = this.configuration.getSettings(context.notebook.uri);
+            const source = [
+                // Prepend cell marker to code
+                context.metadata.interactiveWindowCellMarker ?? settings.defaultCellMarker,
+                context.document.getText()
+            ].join('\n');
+            await this.clipboard.writeText(source);
+        }
+    }
+
+    private getTargetInteractiveWindow(notebookUri: Uri | undefined) {
+        let targetInteractiveWindow;
+        if (notebookUri !== undefined) {
+            targetInteractiveWindow = this.interactiveWindowProvider.windows.find(
+                (w) => w.notebookUri?.toString() === notebookUri.toString()
+            );
+        } else {
+            targetInteractiveWindow = this.interactiveWindowProvider.getActiveOrAssociatedInteractiveWindow();
+        }
+        return targetInteractiveWindow;
     }
 }
