+++ src/interactive-window/interactiveWindow.ts
@@ -72,6 +72,10 @@ import { chainWithPendingUpdates } from '../kernels/execution/notebookUpdater';
 import { initializeInteractiveOrNotebookTelemetryBasedOnUserAction } from '../kernels/telemetry/helper';
 import { generateMarkdownFromCodeLines, parseForComments } from '../platform/common/utils';
 
+/**
+ * ViewModel for an interactive window from the Jupyter extension's point of view.
+ * Methods for talking to an Interactive Window are exposed here, but the actual UI is part of VS code core.
+ */
 export class InteractiveWindow implements IInteractiveWindowLoadable {
     public get onDidChangeViewState(): Event<void> {
         return this._onDidChangeViewState.event;
