+++ src/interactive-window/serviceRegistry.web.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 'use strict';
 
-import { IDataScienceCommandListener } from '../platform/common/types';
 import { ITracebackFormatter } from '../kernels/types';
 import { IExtensionSingleActivationService, IExtensionSyncActivationService } from '../platform/activation/types';
 import { IServiceManager } from '../platform/ioc/types';
@@ -17,7 +16,6 @@ import {
     IDataScienceCodeLensProvider,
     ICodeGeneratorFactory
 } from './editor-integration/types';
-import { InteractiveWindowCommandListener } from './interactiveWindowCommandListener';
 import { InteractiveWindowProvider } from './interactiveWindowProvider';
 import { IInteractiveWindowDebuggingManager, IInteractiveWindowProvider } from './types';
 import { CodeGeneratorFactory } from './editor-integration/codeGeneratorFactory';
@@ -29,10 +27,6 @@ import { InteractiveWindowDebuggingManager } from './debugger/jupyter/debuggingM
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<IInteractiveWindowProvider>(IInteractiveWindowProvider, InteractiveWindowProvider);
-    serviceManager.addSingleton<IDataScienceCommandListener>(
-        IDataScienceCommandListener,
-        InteractiveWindowCommandListener
-    );
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, CommandRegistry);
     serviceManager.add<ICodeWatcher>(ICodeWatcher, CodeWatcher);
     serviceManager.addSingleton<ICodeLensFactory>(ICodeLensFactory, CodeLensFactory);
